Add dynamic import modules logic Add minimal router logic
