+++ src/css/pages/home.scss
@@ -3,7 +3,7 @@
 #home-hero {
   margin-top: -60px;
   width: 100%;
-  min-height: 645px;
+  min-height: 100vh;
   height: 100%;
   background-color: #464646;
   background: url('../../img/bg/pizza-4319999_1280.jpg') rgba(0, 0, 0, 0.5);
@@ -18,7 +18,7 @@
   align-items: center;
 }
 
-#home-about-section-wrapper,
+#home-about-us-section-wrapper,
 #home-menu-section-wrapper,
 #home-whyus-section-wrapper {
   width: 100%;
@@ -49,24 +49,24 @@
   }
 }
 
-#home-about-section {
+#home-about-us-section {
   max-width: 1200px;
   width: 100%;
   height: 509px;
 }
 
 @media only screen and (max-width: 600px) {
-  #home-about-section {
+  #home-about-us-section {
     min-height: 509px;
     height: 100%;
   }
 }
 
-#home-about-section {
+#home-about-us-section {
   display: flex;
   flex-wrap: wrap;
   justify-content: space-between;
-  #home-about-left {
+  #home-about-us-left {
     width: 100%;
     max-width: 600px;
     min-height: 509px;
@@ -90,7 +90,7 @@
     }
   }
 
-  #home-about-right {
+  #home-about-us-right {
     max-width: 482px;
     width: 100%;
     display: flex;
