+++ src/pages/contact.js
@@ -1,0 +1,1 @@
+// R. Itarumirim, numero 48 - Paulo VI, Belo Horizonte - MG, 31998-060, Brezilya
